Common: Rename event_serializer_registry.py -> agent_event_serializer_registry.py
